Extract reducers into their own modules
The root reducer were becoming to big to easily understand and change. This change replaces one
reducer with the redux combineReducers() that only permits reducers to manage a certain slice of
the global state.
